+++ library/md5/src/commonMain/kotlin/org/kotlincrypto/hash/md/MD5.kt
@@ -15,7 +15,7 @@
  **/
 @file:Suppress("UnnecessaryOptInAnnotation")
 
-package org.kotlincrypto.hash
+package org.kotlincrypto.hash.md
 
 import org.kotlincrypto.core.Digest
 import org.kotlincrypto.core.InternalKotlinCryptoApi
