+++ library/sha1/src/commonMain/kotlin/org/kotlincrypto/hash/sha1/SHA1.kt
@@ -15,7 +15,7 @@
  **/
 @file:Suppress("UnnecessaryOptInAnnotation")
 
-package org.kotlincrypto.hash
+package org.kotlincrypto.hash.sha1
 
 import org.kotlincrypto.core.Digest
 import org.kotlincrypto.core.InternalKotlinCryptoApi
