+++ library/sha2/src/commonMain/kotlin/org/kotlincrypto/hash/sha2/SHA256.kt
@@ -13,83 +13,66 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  **/
-package org.kotlincrypto.hash
+package org.kotlincrypto.hash.sha2
 
 import org.kotlincrypto.core.Digest
 import org.kotlincrypto.core.internal.DigestState
 
 /**
- * SHA-384 implementation
+ * SHA-256 implementation
  * */
-public class SHA384: Bit64Digest {
+public class SHA256: Bit32Digest {
 
     public constructor(): super(
-        d = 384,
-        t = null,
-        h0 = -3766243637369397544L,
-        h1 = 7105036623409894663L,
-        h2 = -7973340178411365097L,
-        h3 = 1526699215303891257L,
-        h4 = 7436329637833083697L,
-        h5 = -8163818279084223215L,
-        h6 = -2662702644619276377L,
-        h7 = 5167115440072839076L,
+        d = 256,
+        h0 = 1779033703,
+        h1 = -1150833019,
+        h2 = 1013904242,
+        h3 = -1521486534,
+        h4 = 1359893119,
+        h5 = -1694144372,
+        h6 = 528734635,
+        h7 = 1541459225,
     )
 
-    private constructor(state: DigestState, sha384: SHA384): super(state, sha384)
+    private constructor(state: DigestState, sha256: SHA256): super(state, sha256)
 
-    protected override fun copy(state: DigestState): Digest = SHA384(state, this)
+    protected override fun copy(state: DigestState): Digest = SHA256(state, this)
 
-    protected override fun out(a: Long, b: Long, c: Long, d: Long, e: Long, f: Long, g: Long, h: Long): ByteArray {
+    protected override fun out(a: Int, b: Int, c: Int, d: Int, e: Int, f: Int, g: Int, h: Int): ByteArray {
         return byteArrayOf(
-            (a shr 56).toByte(),
-            (a shr 48).toByte(),
-            (a shr 40).toByte(),
-            (a shr 32).toByte(),
             (a shr 24).toByte(),
             (a shr 16).toByte(),
             (a shr  8).toByte(),
             (a       ).toByte(),
-            (b shr 56).toByte(),
-            (b shr 48).toByte(),
-            (b shr 40).toByte(),
-            (b shr 32).toByte(),
             (b shr 24).toByte(),
             (b shr 16).toByte(),
             (b shr  8).toByte(),
             (b       ).toByte(),
-            (c shr 56).toByte(),
-            (c shr 48).toByte(),
-            (c shr 40).toByte(),
-            (c shr 32).toByte(),
             (c shr 24).toByte(),
             (c shr 16).toByte(),
             (c shr  8).toByte(),
             (c       ).toByte(),
-            (d shr 56).toByte(),
-            (d shr 48).toByte(),
-            (d shr 40).toByte(),
-            (d shr 32).toByte(),
             (d shr 24).toByte(),
             (d shr 16).toByte(),
             (d shr  8).toByte(),
             (d       ).toByte(),
-            (e shr 56).toByte(),
-            (e shr 48).toByte(),
-            (e shr 40).toByte(),
-            (e shr 32).toByte(),
             (e shr 24).toByte(),
             (e shr 16).toByte(),
             (e shr  8).toByte(),
             (e       ).toByte(),
-            (f shr 56).toByte(),
-            (f shr 48).toByte(),
-            (f shr 40).toByte(),
-            (f shr 32).toByte(),
             (f shr 24).toByte(),
             (f shr 16).toByte(),
             (f shr  8).toByte(),
-            (f       ).toByte()
+            (f       ).toByte(),
+            (g shr 24).toByte(),
+            (g shr 16).toByte(),
+            (g shr  8).toByte(),
+            (g       ).toByte(),
+            (h shr 24).toByte(),
+            (h shr 16).toByte(),
+            (h shr  8).toByte(),
+            (h       ).toByte()
         )
     }
 }
