+++ library/sha2/src/commonMain/kotlin/org/kotlincrypto/hash/sha2/SHA512.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  **/
-package org.kotlincrypto.hash
+package org.kotlincrypto.hash.sha2
 
 import org.kotlincrypto.core.Digest
 import org.kotlincrypto.core.internal.DigestState
