+++ library/sha2/src/commonMain/kotlin/org/kotlincrypto/hash/sha2/SHA512t.kt
@@ -15,7 +15,7 @@
  **/
 @file:Suppress("FunctionName")
 
-package org.kotlincrypto.hash
+package org.kotlincrypto.hash.sha2
 
 import org.kotlincrypto.core.Digest
 import org.kotlincrypto.core.internal.DigestState
