+++ library/sha2/src/commonTest/kotlin/org/kotlincrypto/hash/sha2/SHA224UnitTest.kt
@@ -13,9 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  **/
-package org.kotlincrypto.hash
+package org.kotlincrypto.hash.sha2
 
 import org.kotlincrypto.core.Digest
+import org.kotlincrypto.hash.DigestUnitTest
 import kotlin.test.Test
 
 open class SHA224UnitTest: DigestUnitTest() {
