+++ library/sha2/src/jvmTest/kotlin/org/kotlincrypto/hash/sha2/SHA224JvmUnitTest.kt
@@ -13,10 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  **/
-package org.kotlincrypto.hash
+package org.kotlincrypto.hash.sha2
 
 import org.kotlincrypto.core.Digest
+import org.kotlincrypto.hash.TestJvmDigest
 
-class SHA512JvmUnitTest: SHA512UnitTest() {
+class SHA224JvmUnitTest: SHA224UnitTest() {
     override val digest: Digest = TestJvmDigest(super.digest)
 }
