+++ library/sha2/src/jvmTest/kotlin/org/kotlincrypto/hash/sha2/SHA256JvmUnitTest.kt
@@ -13,10 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  **/
-package org.kotlincrypto.hash
+package org.kotlincrypto.hash.sha2
 
 import org.kotlincrypto.core.Digest
+import org.kotlincrypto.hash.TestJvmDigest
 
-class MD5JvmUnitTest: MD5UnitTest() {
+class SHA256JvmUnitTest: SHA256UnitTest() {
     override val digest: Digest = TestJvmDigest(super.digest)
 }
