+++ library/sha2/src/jvmTest/kotlin/org/kotlincrypto/hash/sha2/SHA384JvmUnitTest.kt
@@ -13,10 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  **/
-package org.kotlincrypto.hash
+package org.kotlincrypto.hash.sha2
 
 import org.kotlincrypto.core.Digest
+import org.kotlincrypto.hash.TestJvmDigest
 
-class SHA256JvmUnitTest: SHA256UnitTest() {
+class SHA384JvmUnitTest: SHA384UnitTest() {
     override val digest: Digest = TestJvmDigest(super.digest)
 }
