+++ library/sha2/src/jvmTest/kotlin/org/kotlincrypto/hash/sha2/SHA512_224JvmUnitTest.kt
@@ -13,11 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  **/
-package org.kotlincrypto.hash
+package org.kotlincrypto.hash.sha2
 
 import org.kotlincrypto.core.Digest
+import org.kotlincrypto.hash.TestJvmDigest
 
 @Suppress("ClassName")
-class SHA512_256Jvm256UnitTest: SHA512_256UnitTest() {
+class SHA512_224JvmUnitTest: SHA512_224UnitTest() {
     override val digest: Digest = TestJvmDigest(super.digest)
 }
